refactor: split large fill-stateful file
