+++ packages/testing/src/execution_testing/cli/pytest_commands/plugins/fill_stateful/fill_stateful.py
@@ -37,30 +37,20 @@ from execution_testing.fixtures.blockchain import (
 )
 from execution_testing.forks import (
     Fork,
-    ForkSetAdapter,
-    InvalidForkError,
     TransitionFork,
 )
 from execution_testing.logging import get_logger
-from execution_testing.rpc import EngineRPC, EthRPC
+from execution_testing.rpc import EthRPC
 from execution_testing.specs.blockchain import (
     payload_metadata_to_fixture,
 )
 from execution_testing.test_types import EOA
-from execution_testing.test_types.chain_config_types import (
-    ChainConfigDefaults,
-)
 
 from ..execute import contracts
 from ..execute.rpc.chain_builder_eth_rpc import ChainBuilderEthRPC
-from ..execute.rpc.hive import (
-    build_client_files,
-    build_client_genesis_dict,
-    build_genesis_header,
-    build_hive_environment,
-)
 from ..shared.helpers import is_help_or_collectonly_mode
 from ..shared.live_client_flags import FEE_BUMP_MULTIPLIER
+from .hive_session import configure_hive, teardown_hive
 
 # 1B ETH for seed account
 SEED_FUNDING_WEI = 10**9 * 10**18
@@ -163,122 +153,6 @@ def pytest_addoption(parser: pytest.Parser) -> None:
     )
 
 
-def _resolve_session_fork(
-    config: pytest.Config,
-) -> Fork | TransitionFork:
-    """
-    Resolve the single fork from ``--fork`` in hive mode.
-
-    Runs in ``pytest_configure`` before forks.py populates
-    ``selected_fork_set``, so we can't use ``session_fork`` here — the
-    hive genesis/ruleset must be built before remote.py's pytest_configure
-    pings the (not-yet-started) client.
-    """
-    fork_arg = config.getoption("single_fork", default="")
-    if not fork_arg:
-        pytest.exit(
-            "--fork is required in --hive-mode (single-fork only).",
-            returncode=pytest.ExitCode.USAGE_ERROR,
-        )
-    try:
-        fork_set = ForkSetAdapter.validate_python(fork_arg)
-    except InvalidForkError:
-        pytest.exit(
-            f"Unsupported fork provided to --fork: {fork_arg!r}",
-            returncode=pytest.ExitCode.USAGE_ERROR,
-        )
-    if len(fork_set) != 1:
-        pytest.exit(
-            f"Expected exactly one fork in --hive-mode, got {len(fork_set)} "
-            f"({sorted(f.name() for f in fork_set)}).",
-            returncode=pytest.ExitCode.USAGE_ERROR,
-        )
-    return next(iter(fork_set))
-
-
-def _hive_chain_id(config: pytest.Config) -> int:
-    """Resolve chain id from CLI/env, defaulting to ChainConfigDefaults."""
-    cli_value = config.getoption("chain_id", default=None)
-    if cli_value is not None:
-        return int(cli_value)
-    env_value = os.environ.get("CHAIN_ID")
-    if env_value is not None:
-        return int(env_value)
-    return ChainConfigDefaults.chain_id
-
-
-def _configure_hive(config: pytest.Config) -> None:
-    """
-    Start a hive simulator session and a single execution client, then
-    rewrite ``config.option`` so the rest of the plugin stack (remote.py
-    in particular) connects to the hive-managed client.
-    """
-    hive_simulator_url = config.getoption("hive_simulator")
-    if hive_simulator_url is None:
-        pytest.exit(
-            "The HIVE_SIMULATOR environment variable is not set.\n\n"
-            "If running locally, start hive in --dev mode, for example:\n"
-            "./hive --dev --client go-ethereum\n\n"
-            "and set the HIVE_SIMULATOR to the reported URL. For example, "
-            "in bash:\n"
-            "export HIVE_SIMULATOR=http://127.0.0.1:3000\n"
-            "or in fish:\n"
-            "set -x HIVE_SIMULATOR http://127.0.0.1:3000"
-        )
-    from hive.simulation import Simulation
-
-    session_fork = _resolve_session_fork(config)
-    chain_id = _hive_chain_id(config)
-
-    simulator = Simulation(url=hive_simulator_url)
-    suite = simulator.start_suite(
-        name="fill-stateful test suite",
-        description="Test suite used to drive a fill-stateful session",
-    )
-    config.hive_test_suite = suite  # type: ignore[attr-defined]
-    base_test = suite.start_test(
-        name="fill-stateful base test",
-        description=(
-            "Base test in the fill-stateful suite hosting the long-lived "
-            "execution client."
-        ),
-    )
-    config.hive_base_test = base_test  # type: ignore[attr-defined]
-
-    # base_pre=None: seed key funded via CL withdrawal and deterministic
-    # factory deployed on-chain by ``_session_pre_run``, so genesis only
-    # needs the fork's required system contracts.
-    pre_alloc, genesis_header = build_genesis_header(session_fork, None)
-    genesis_dict = build_client_genesis_dict(pre_alloc, genesis_header)
-
-    client_type = simulator.client_types()[0]
-    client = base_test.start_client(
-        client_type=client_type,
-        environment=build_hive_environment(session_fork, chain_id),
-        files=build_client_files(genesis_dict),
-    )
-    if client is None:
-        pytest.exit(
-            f"Unable to start hive client {client_type.name}. Check the "
-            "hive server logs for more information."
-        )
-    config.hive_client = client  # type: ignore[attr-defined]
-    logger.info(
-        f"Started hive client {client_type.name} at {client.ip} "
-        f"(fork={session_fork.name()}, chain_id={chain_id})"
-    )
-
-    config.option.rpc_endpoint = f"http://{client.ip}:8545"
-    config.option.engine_endpoint = f"http://{client.ip}:8551"
-    if (
-        config.getoption("engine_jwt_secret", default=None) is None
-        and config.getoption("engine_jwt_secret_file", default=None) is None
-    ):
-        config.option.engine_jwt_secret = EngineRPC.DEFAULT_JWT_SECRET
-    if config.getoption("chain_id", default=None) is None:
-        config.option.chain_id = chain_id
-
-
 @pytest.hookimpl(tryfirst=True)
 def pytest_configure(config: pytest.Config) -> None:
     """
@@ -310,7 +184,7 @@ def pytest_configure(config: pytest.Config) -> None:
     config.hive_base_test = None  # type: ignore[attr-defined]
     config.hive_client = None  # type: ignore[attr-defined]
     if config.getoption("hive_mode", default=False):
-        _configure_hive(config)
+        configure_hive(config)
     else:
         if not config.getoption("rpc_endpoint", default=None):
             config.option.rpc_endpoint = "http://localhost:8545"
@@ -328,43 +202,7 @@ def pytest_configure(config: pytest.Config) -> None:
 
 def pytest_unconfigure(config: pytest.Config) -> None:
     """Tear down hive resources started in ``pytest_configure``."""
-    client = getattr(config, "hive_client", None)
-    if client is not None:
-        try:
-            client.stop()
-        except Exception as e:
-            logger.warning(f"Failed to stop hive client: {e}")
-        config.hive_client = None  # type: ignore[attr-defined]
-
-    base_test = getattr(config, "hive_base_test", None)
-    if base_test is not None:
-        from hive.testing import HiveTestResult
-
-        test_pass = (
-            getattr(config, "_fill_stateful_session_failed", False) is False
-        )
-        try:
-            base_test.end(
-                result=HiveTestResult(
-                    test_pass=test_pass,
-                    details=(
-                        "fill-stateful session complete"
-                        if test_pass
-                        else "fill-stateful session had failures"
-                    ),
-                )
-            )
-        except Exception as e:
-            logger.warning(f"Failed to end hive base test: {e}")
-        config.hive_base_test = None  # type: ignore[attr-defined]
-
-    suite = getattr(config, "hive_test_suite", None)
-    if suite is not None:
-        try:
-            suite.end()
-        except Exception as e:
-            logger.warning(f"Failed to end hive test suite: {e}")
-        config.hive_test_suite = None  # type: ignore[attr-defined]
+    teardown_hive(config)
 
 
 def pytest_sessionfinish(session: pytest.Session, exitstatus: int) -> None:
@@ -461,12 +299,6 @@ def max_gas_limit_per_test(
         return cli_value
     fork_at_genesis = session_fork.fork_at(block_number=0, timestamp=0)
     return fork_at_genesis.transaction_gas_limit_cap()
-
-
-# Other live-client fixtures (``max_transactions_per_batch``,
-# ``default_*``, fee fields, ``dry_run``, ...) come from
-# ``shared.live_client_flags``. ``skip_cleanup`` from ``execute.pre_alloc``;
-# we force it on in ``pytest_configure``.
 
 
 # ---------------------------------------------------------------------------
